refactor: move queue managing from MusicService to own class
