+++ app/src/main/java/com/dkanada/gramophone/helper/MusicPlayerRemote.java
@@ -14,6 +14,7 @@ import android.widget.Toast;
 import com.dkanada.gramophone.R;
 import com.dkanada.gramophone.model.Song;
 import com.dkanada.gramophone.service.MusicService;
+import com.dkanada.gramophone.service.QueueManager;
 import com.dkanada.gramophone.util.PreferenceUtil;
 
 import java.util.ArrayList;
@@ -150,7 +151,7 @@ public class MusicPlayerRemote {
         if (!tryToHandleOpenPlayingQueue(queue, startPosition, startPlaying) && musicService != null) {
             musicService.openQueue(queue, startPosition, startPlaying);
             if (!PreferenceUtil.getInstance(musicService).getRememberShuffle()){
-                setShuffleMode(MusicService.SHUFFLE_MODE_NONE);
+                setShuffleMode(QueueManager.SHUFFLE_MODE_NONE);
             }
         }
     }
@@ -163,7 +164,7 @@ public class MusicPlayerRemote {
 
         if (!tryToHandleOpenPlayingQueue(queue, startPosition, startPlaying) && musicService != null) {
             openQueue(queue, startPosition, startPlaying);
-            setShuffleMode(MusicService.SHUFFLE_MODE_SHUFFLE);
+            setShuffleMode(QueueManager.SHUFFLE_MODE_SHUFFLE);
         }
     }
 
@@ -182,24 +183,24 @@ public class MusicPlayerRemote {
     }
 
     public static Song getCurrentSong() {
-        if (musicService != null) {
-            return musicService.getCurrentSong();
+        if (musicService != null && musicService.queueManager != null) {
+            return musicService.queueManager.getCurrentSong();
         }
 
         return null;
     }
 
     public static int getPosition() {
-        if (musicService != null) {
-            return musicService.getPosition();
+        if (musicService != null && musicService.queueManager != null) {
+            return musicService.queueManager.getPosition();
         }
 
         return -1;
     }
 
     public static List<Song> getPlayingQueue() {
-        if (musicService != null) {
-            return musicService.getPlayingQueue();
+        if (musicService != null && musicService.queueManager != null) {
+            return musicService.queueManager.getPlayingQueue();
         }
 
         return new ArrayList<>();
@@ -222,8 +223,8 @@ public class MusicPlayerRemote {
     }
 
     public static long getQueueDurationMillis(int position) {
-        if (musicService != null) {
-            return musicService.getQueueDurationMillis(position);
+        if (musicService != null && musicService.queueManager != null) {
+            return musicService.queueManager.getQueueDurationMillis(position);
         }
 
         return -1;
@@ -238,24 +239,24 @@ public class MusicPlayerRemote {
     }
 
     public static int getRepeatMode() {
-        if (musicService != null) {
-            return musicService.getRepeatMode();
+        if (musicService != null && musicService.queueManager != null) {
+            return musicService.queueManager.getRepeatMode();
         }
 
-        return MusicService.REPEAT_MODE_NONE;
+        return QueueManager.REPEAT_MODE_NONE;
     }
 
     public static int getShuffleMode() {
-        if (musicService != null) {
-            return musicService.getShuffleMode();
+        if (musicService != null && musicService.queueManager != null) {
+            return musicService.queueManager.getShuffleMode();
         }
 
-        return MusicService.SHUFFLE_MODE_NONE;
+        return QueueManager.SHUFFLE_MODE_NONE;
     }
 
     public static boolean cycleRepeatMode() {
-        if (musicService != null) {
-            musicService.cycleRepeatMode();
+        if (musicService != null && musicService.queueManager != null) {
+            musicService.queueManager.cycleRepeatMode();
             return true;
         }
 
@@ -263,8 +264,8 @@ public class MusicPlayerRemote {
     }
 
     public static boolean toggleShuffleMode() {
-        if (musicService != null) {
-            musicService.toggleShuffle();
+        if (musicService != null && musicService.queueManager != null) {
+            musicService.queueManager.toggleShuffle();
             return true;
         }
 
@@ -272,8 +273,8 @@ public class MusicPlayerRemote {
     }
 
     public static boolean setShuffleMode(final int shuffleMode) {
-        if (musicService != null) {
-            musicService.setShuffleMode(shuffleMode);
+        if (musicService != null && musicService.queueManager != null) {
+            musicService.queueManager.setShuffleMode(shuffleMode);
             return true;
         }
 
@@ -281,9 +282,9 @@ public class MusicPlayerRemote {
     }
 
     public static boolean playNext(Song song) {
-        if (musicService != null) {
+        if (musicService != null && musicService.queueManager != null) {
             if (getPlayingQueue().size() > 0) {
-                musicService.addSong(getPosition() + 1, song);
+                musicService.queueManager.addSong(getPosition() + 1, song);
             } else {
                 List<Song> queue = new ArrayList<>();
                 queue.add(song);
@@ -298,9 +299,9 @@ public class MusicPlayerRemote {
     }
 
     public static boolean playNext(@NonNull List<Song> songs) {
-        if (musicService != null) {
+        if (musicService != null && musicService.queueManager != null) {
             if (getPlayingQueue().size() > 0) {
-                musicService.addSongs(getPosition() + 1, songs);
+                musicService.queueManager.addSongs(getPosition() + 1, songs);
             } else {
                 openQueue(songs, 0, false);
             }
@@ -314,9 +315,9 @@ public class MusicPlayerRemote {
     }
 
     public static boolean enqueue(Song song) {
-        if (musicService != null) {
+        if (musicService != null && musicService.queueManager != null) {
             if (getPlayingQueue().size() > 0) {
-                musicService.addSong(song);
+                musicService.queueManager.addSong(song);
             } else {
                 List<Song> queue = new ArrayList<>();
                 queue.add(song);
@@ -331,9 +332,9 @@ public class MusicPlayerRemote {
     }
 
     public static boolean enqueue(@NonNull List<Song> songs) {
-        if (musicService != null) {
+        if (musicService != null && musicService.queueManager != null) {
             if (getPlayingQueue().size() > 0) {
-                musicService.addSongs(songs);
+                musicService.queueManager.addSongs(songs);
             } else {
                 openQueue(songs, 0, false);
             }
@@ -347,8 +348,8 @@ public class MusicPlayerRemote {
     }
 
     public static boolean removeFromQueue(int position) {
-        if (musicService != null && position >= 0 && position < getPlayingQueue().size()) {
-            musicService.removeSong(position);
+        if (musicService != null && musicService.queueManager != null && position >= 0 && position < getPlayingQueue().size()) {
+            musicService.queueManager.removeSong(position);
             return true;
         }
 
@@ -356,8 +357,8 @@ public class MusicPlayerRemote {
     }
 
     public static boolean moveSong(int from, int to) {
-        if (musicService != null && from >= 0 && to >= 0 && from < getPlayingQueue().size() && to < getPlayingQueue().size()) {
-            musicService.moveSong(from, to);
+        if (musicService != null && musicService.queueManager != null && from >= 0 && to >= 0 && from < getPlayingQueue().size() && to < getPlayingQueue().size()) {
+            musicService.queueManager.moveSong(from, to);
             return true;
         }
 
@@ -365,8 +366,8 @@ public class MusicPlayerRemote {
     }
 
     public static boolean clearQueue() {
-        if (musicService != null) {
-            musicService.clearQueue();
+        if (musicService != null && musicService.queueManager != null) {
+            musicService.queueManager.clearQueue();
             return true;
         }
