+++ app/src/main/java/com/dkanada/gramophone/views/widgets/AppWidgetAlbum.java
@@ -68,7 +68,7 @@ public class AppWidgetAlbum extends BaseAppWidget {
         final RemoteViews appWidgetView = new RemoteViews(service.getPackageName(), R.layout.app_widget_album);
 
         final boolean isPlaying = service.isPlaying();
-        final Song song = service.getCurrentSong();
+        final Song song = service.queueManager.getCurrentSong();
 
         if (TextUtils.isEmpty(song.title) && TextUtils.isEmpty(song.artistName)) {
             appWidgetView.setViewVisibility(R.id.media_titles, View.INVISIBLE);
