+++ app/src/main/java/com/dkanada/gramophone/views/widgets/AppWidgetClassic.java
@@ -66,7 +66,7 @@ public class AppWidgetClassic extends BaseAppWidget {
         final RemoteViews appWidgetView = new RemoteViews(service.getPackageName(), R.layout.app_widget_classic);
 
         final boolean isPlaying = service.isPlaying();
-        final Song song = service.getCurrentSong();
+        final Song song = service.queueManager.getCurrentSong();
 
         if (TextUtils.isEmpty(song.title) && TextUtils.isEmpty(song.artistName)) {
             appWidgetView.setViewVisibility(R.id.media_titles, View.INVISIBLE);
